Refactor assertValidJSON into a shared mixin utility
